feat: Add extra password error handling
